+++ src/amr/sync/utils.js
@@ -1,3 +1,0 @@
-export const DELETED = 1
-
-export const FAIL_KEY = "_no_key_"